+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 Version 0.2.3 (currently in development)
 ========================================
 
+* Added Configuration Cache
+
 Version 0.2.2 (2015-02-08)
 ========================================